deleteItems,moveItems will refresh parent shelf from which move/delete is happening using onTransition state function.
pending : refreshing invalidated shelf which user come to it via going up the tree.

solution to above pending part :
on folder open we replace the screen with new screen...now when we are destroying/replacing new screen(open shelf) we check if the array of previous shelfId (going up the tree [poping screen]) is empty or not ..if not empty we take last id and replace screen and on init we check if the shelf id is invalid or not if invalid we fetch from page 1./current shelfId is invalid or not if invalid we fetch from pageNo1..

summary is check current shelf Id is in invalidate state or not if invalid fetch from pageNo 1.

the case where items are delted from shelf (neight poping nor pushing) has been handled using onTransition.